requirements.txt
apache-flink==1.18.1
ibis-framework[postgres, duckdb, risingwave, datafusion, polars]>=8
kafka-python
numpy==1.24.4
pins
gcsfs
dill